biovault-fixed/biovault-fixed/voice-service/requirements.txt
fastapi
uvicorn
torch
torchaudio
speechbrain
openai-whisper
python-multipart
librosa